Task 6 Step 2: Create and export Postman collection for httpbin.org requests
