--- src/main/java/library/config/AppConfig.java
@@ -1,77 +1,0 @@
-package library.config;
-
-
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.MessageSource;
-import org.springframework.context.annotation.EnableAspectJAutoProxy;
-import org.springframework.context.support.ResourceBundleMessageSource;
-
-import java.io.IOException;
-import java.io.InputStream;
-import java.util.Properties;
-import org.hibernate.SessionFactory;
-import org.springframework.orm.hibernate5.LocalSessionFactoryBean;
-import org.springframework.jdbc.datasource.DriverManagerDataSource;
-import javax.sql.DataSource;
-
-@Configuration
-@ComponentScan(basePackages = "library")
-@EnableAspectJAutoProxy
-public class AppConfig {
-    @Bean
-    public MessageSource messageSource(){
-        ResourceBundleMessageSource messageSource = new ResourceBundleMessageSource();
-        messageSource.setBasename("messages");
-        messageSource.setDefaultEncoding("UTF-8");
-        return messageSource;
-    }
-    @Bean
-    public static DataSource dataSource() throws ClassNotFoundException {
-        Properties props = loadConfigProperties();
-        DriverManagerDataSource dataSource = new DriverManagerDataSource();
-        dataSource.setDriverClassName(props.getProperty("db.driver"));
-        dataSource.setUrl(props.getProperty("db.url"));
-        dataSource.setUsername(props.getProperty("db.username"));
-        dataSource.setPassword(props.getProperty("db.password"));
-        return dataSource;
-    }
-    private static Properties loadConfigProperties() throws ClassNotFoundException {
-        Properties props = new Properties();
-        try(InputStream inputStream = AppConfig.class.getClassLoader().getResourceAsStream("application-local.properties")){
-            props.load(inputStream);
-        } catch (IOException e) {
-            throw new ClassNotFoundException("Failed to load the config details",e);
-        }
-        return props;
-
-    }
-
-
-    @Bean
-    public LocalSessionFactoryBean sessionFactory(DataSource dataSource) {
-        LocalSessionFactoryBean sessionFactory = new LocalSessionFactoryBean();
-        sessionFactory.setDataSource(dataSource);
-        sessionFactory.setPackagesToScan("library.model");
-        Properties hibernateProperties = new Properties();
-        hibernateProperties.setProperty("hibernate.hbm2ddl.auto", "validate");
-        hibernateProperties.setProperty("hibernate.show_sql", "true");
-        hibernateProperties.setProperty("hibernate.format_sql", "true");
-        hibernateProperties.setProperty("hibernate.cache.use_second_level_cache", "true");
-        hibernateProperties.setProperty("hibernate.cache.use_query_cache", "true");
-        hibernateProperties.setProperty("hibernate.cache.region.factory_class", "org.hibernate.cache.jcache.internal.JCacheRegionFactory");
-        hibernateProperties.setProperty("hibernate.jdbc.batch_size", "50");
-        hibernateProperties.setProperty("hibernate.order_inserts", "true");
-        hibernateProperties.setProperty("hibernate.order_updates", "true");
-        hibernateProperties.setProperty("hibernate.javax.cache.provider","org.ehcache.jsr107.EhcacheCachingProvider");
-       hibernateProperties.setProperty("hibernate.javax.cache.missing_cache_strategy" , "create");
-        sessionFactory.setHibernateProperties(hibernateProperties);
-        return sessionFactory;
-    }
-
-    @Bean
-    public SessionFactory sessionFactoryBean(LocalSessionFactoryBean sessionFactory) throws Exception {
-        return sessionFactory.getObject();
-    }
-}
