--- src/main/resources/application-local.properties
@@ -1,4 +1,0 @@
-db.driver=org.postgresql.Driver
-db.url=jdbc:postgresql://localhost:5440/Library
-db.username=bereketab
-db.password=bereketabInnowise
