+++ src/main/resources/application.properties
@@ -1,6 +1,27 @@
-#Please fill out your db credentials here
-db.driver=${DB_DRIVER}
-db.url=${DB_URL}
-db.username=${DB_USERNAME}
-db.password=${DB_PASSWORD}
 
+# Database
+spring.datasource.url=jdbc:postgresql://localhost:5440/Library
+spring.datasource.username=bereketab
+spring.datasource.password=bereketabInnowise
+spring.datasource.driver-class-name=org.postgresql.Driver
+
+# Hibernate
+spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.PostgreSQLDialect
+spring.jpa.hibernate.ddl-auto=validate
+spring.jpa.show-sql=true
+spring.jpa.properties.hibernate.format_sql=true
+
+# Second-level cache
+spring.jpa.properties.hibernate.cache.use_second_level_cache=true
+spring.jpa.properties.hibernate.cache.use_query_cache=true
+spring.jpa.properties.hibernate.cache.region.factory_class=org.hibernate.cache.jcache.JCacheRegionFactory
+spring.jpa.properties.hibernate.javax.cache.provider=org.ehcache.jsr107.EhcacheCachingProvider
+spring.jpa.properties.hibernate.javax.cache.missing_cache_strategy=create
+# Pool
+spring.datasource.hikari.maximum-pool-size=10
+# Liquibase
+spring.liquibase.enabled=false
+# Server
+server.port=8080
+# Logback
+logging.level.ch.qos.logback=DEBUG
